First pass at organizing models
