+++ src/context/PagesContext.tsx
@@ -1,7 +1,7 @@
 import React, { createContext, useEffect, useState } from "react";
 import { getPages } from "../fetchers";
 import { OK } from "node-kall";
-import { PageModel } from "../models";
+import { PageModel } from "../models/models";
