+++ src/fetchers.ts
@@ -1,5 +1,5 @@
 import { get, put, del, post } from "node-kall";
-import { PageModel, ResponseModel, CategoryModel, EmailModel, PaymentRequestModel, CoordinateModel } from "./models";
+import { PageModel, ResponseModel, CategoryModel, EmailModel, PaymentRequestModel, CoordinateModel } from "./models/models";
 import Stripe from "stripe";
 
 export const putPage = (page: PageModel) =>
