+++ src/models/AuthModel.ts
@@ -1,0 +1,6 @@
+//from Auth0
+export interface AuthModel {
+    name: string,
+    sub: string,
+    email: string
+}
