+++ src/models/CategoryModel.ts
@@ -1,0 +1,6 @@
+export interface CategoryModel {
+    id?: string,
+    name: string
+    owner_id: string,
+    created_at?: string,
+}
