+++ src/models/CoordinateModel.ts
@@ -1,0 +1,6 @@
+// Charts 
+export interface CoordinateModel {
+    y: number | string,
+    x: number | string | Date,
+    label?: string
+}
