+++ src/models/EmailModel.ts
@@ -1,0 +1,5 @@
+export interface EmailModel {
+    to?: string,
+    from?: string,
+    text: string
+}
