+++ src/models/Emotion.ts
@@ -1,0 +1,2 @@
+export type Emotion =
+    ':-)' | ':-|' | ':-('
