+++ src/models/PageModel.ts
@@ -1,0 +1,8 @@
+export interface PageModel {
+    id: string,
+    name: string,
+    owner_id: string,
+    created_at?: string,
+    color?: string,
+    category_id?: string
+}
