+++ src/models/PaymentModel.ts
@@ -1,0 +1,6 @@
+// Payment API 
+export interface PaymentRequestModel {
+    customerId: string,
+    paymentMethodId: string,
+    priceId: string
+};
