+++ src/pages/api/pages/[id].ts
@@ -2,7 +2,7 @@ import auth0 from "../../../auth/auth0";
 import { pages } from "../../../database/database";
 import { NOT_FOUND, BAD_REQUEST, UNAUTHORIZED, NOT_IMPLEMENTED, FORBIDDEN, NO_CONTENT } from "node-kall";
 import { NextApiRequest, NextApiResponse } from "next";
-import { PageModel } from "../../../models";
+import { PageModel } from "../../../models/models";
 import { withCors } from "../../../middleware/middleware";
 import { getId } from "../users/[id]";
