+++ tests/api/pages/id.test.ts
@@ -4,7 +4,7 @@ import { users, categories, pages } from "../../../src/database/database"
 import { randomUser } from "../../database/databaseTestUtils";
 import { setupServer, teardownServer, uid, authenticatedFetch, randomPage } from "../apiTestUtils";
 import handler from "../../../src/pages/api/pages/[id]";
-import { PageModel } from "../../../src/models";
+import { PageModel } from "../../../src/models/models";
