+++ tests/database/categories.test.ts
@@ -1,7 +1,7 @@
 import * as faker from "faker";
 import { categories } from "../../src/database/categories";
 import { randomUser } from "./databaseTestUtils";
-import { CategoryModel } from "../../src/models";
+import { CategoryModel } from "../../src/models/models";
 import { users } from "../../src/database/users";
 
 describe("Database interface for categories", () => {
