+++ tests/frontend/admin/admin.test.tsx
@@ -6,7 +6,7 @@ import React from "react";
 import { AdminPage } from "../../../src/components/Admin/AdminPage"
 import { TextList } from "../../../src/components/Admin/TextList/TextList";
 import { waitFor, render } from "@testing-library/react"
-import { Emotion, ResponseModel } from "../../../src/models";
+import { Emotion, ResponseModel } from "../../../src/models/models";
 import * as text from "../../../src/text"
 import '@testing-library/jest-dom/extend-expect'
 import * as faker from "faker";
