+++ tests/frontend/page/ResponseSection.test.tsx
@@ -6,7 +6,7 @@
 import { waitFor, fireEvent, render } from "@testing-library/react"
 import '@testing-library/jest-dom/extend-expect'
 import { ResponseSection } from "../../../src/components/Page/ResponseSection"
-import { PageModel } from "../../../src/models";
+import { PageModel } from "../../../src/models/models";
 import { randomPage, randomUser } from "../../database/databaseTestUtils";
 
 describe("The component for creating new responses", () => {
